Performance improvement: avoid storing event matchings in database
This comes at a slight expense of having an inaccurate total_matching
count — I’m working on a fix to make it a bit more accurate. It will
never be 100% accurate without storing event matchings, but it should
be good enough, and the performance gainz are more than worth it.